% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pathnames.R
\name{filter_pathnames}
\alias{filter_pathnames}
\title{filter directory names by regular expressions}
\usage{
filter_pathnames(search_regexes, paths, reject_regex = NULL)
}
\arguments{
\item{search_regexes}{vector of regexes to filter down the paths you are searching}

\item{paths}{character vector of directory paths}

\item{reject_regex}{single regex string to reject any filenames that match the regex (broaden with regex 'or' pipes '|')}
}
\value{
path names of the pdf documents that satisfy your regexes
}
\description{
(used/called inside pdf_search_txt)
}
\examples{
\dontrun{
  regexes <- c('(?i)wire', '\\\\.pdf')
  filter_pathnames(regexes, list.files(here::here(), recursive = T))
 }
}
