% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pdf_contents.R
\name{filter_pdf_contents}
\alias{filter_pdf_contents}
\title{filter pdf documents by search via regular expressions}
\usage{
filter_pdf_contents(search_regexes, pdf_contents, reject_regex = NULL)
}
\arguments{
\item{search_regexes}{vector of regexes to filter down the pdfs you're searching through}

\item{pdf_contents}{the list of pdfs that have been read as text either by pdftools::pdf_text or ocr}

\item{reject_regex}{single regex to reject any files that match the regex (broaden with regex 'or' pipes '|')}
}
\value{
path names of the pdf documents that satisfy your regexes
}
\description{
(used/called inside pdf_search_txt)
}
\examples{
 \dontrun{ filter_pdf_contents(c('(?i)blue','Green'), contents) }

}
