% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_embed_ocr.R
\name{pdf_embed_ocr}
\alias{pdf_embed_ocr}
\title{OCR a pdf and embed text
R implementation of OCRmyPDF for linux
To install tesseract:
https://tesseract-ocr.github.io/tessdoc/Installation.html
https://github.com/UB-Mannheim/tesseract/wiki
https://ocrmypdf.readthedocs.io/en/latest/installation.html#installing-on-windows
To install on Linux https://ocrmypdf.readthedocs.io/en/latest/installation.html}
\usage{
pdf_embed_ocr(path_read, ..., path_save = NULL)
}
\arguments{
\item{path_read}{text: path to the pdf to ocr (uses here)}

\item{...}{text: optional args for OCRmyPDF (https://ocrmypdf.readthedocs.io/en/latest/cookbook.html#)}

\item{path_save}{text: (optional) save location/filename.pdf (if used, argument must be specified by name)}
}
\value{
a path to the OCR'd output pdf file
}
\description{
To see all options, in terminal type: man ocrmypdf
}
\examples{
\dontrun{
pdf_embed_ocr('/home/user/Desktop/somefile.pdf','--deskew', '--clean')
list.files(pattern = '(?i)pdf$', full.names = T, recursive = T) |> #' purrr::map(
~pdf_embed_ocr(., '--deskew', '--rotate-pages', '--threshold', '--skip-text',
'--redo-ocr', '--clean','-remove-background'))
}

}
