% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_insert.R
\name{pdf_insert}
\alias{pdf_insert}
\title{insert one PDF into another at a page location}
\usage{
pdf_insert(
  main_pdf = "",
  insert_pdf = "",
  after_page = 0,
  remove_main_pages = c(0),
  remove_inserted_pages = c(0),
  output = NULL,
  password = ""
)
}
\arguments{
\item{main_pdf}{string: the primary pdf that takes the insert}

\item{insert_pdf}{string: the pdf to insert into main_pdf}

\item{after_page}{integer: page number (if no entry, inserted pdf will pre-pend)}

\item{remove_main_pages}{integers: vector of pages from the primary pdf document to remove}

\item{remove_inserted_pages}{integers: vector of pages from the insert pdf document to remove}

\item{output}{string: output pdf name or path with output pdf name}

\item{password}{string: password to the pdfs if there is one}
}
\value{
character path to the output pdf
}
\description{
insert one PDF into another at a page location
}
\examples{
\dontrun{
pdf_insert('folder/main_document.pdf', 'folder/insert_this.pdf', after_page = 10)
}
}
