% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_search_txt.R
\name{pdf_search_txt}
\alias{pdf_search_txt}
\title{Search folder for PDF's whose content matches your regular expression}
\usage{
pdf_search_txt(
  search_regexes = c("", ""),
  search_folder = here::here(),
  recursive = FALSE,
  reject_regex = NULL,
  include_pathnames = TRUE,
  reject_pathnames_regex = NULL,
  ignore_duplicate_filenames = FALSE,
  output_form = c("doc_paths", "doc_text", "matching_pages", "text_lines"),
  reread = FALSE
)
}
\arguments{
\item{search_regexes}{vector of regular expressions to filter down the pdfs you're searching through}

\item{search_folder}{a folder location that contains the pdfs you want to read (uses here())}

\item{recursive}{search in specified folder as well as its sub-folders}

\item{reject_regex}{single regex to reject any files that match the regex (broaden with stringr::regex() 'or' pipes '|')}

\item{include_pathnames}{search for the regex text in pathnames as well}

\item{reject_pathnames_regex}{single regex to reject any pathnames that match (broaden with stringr::regex() 'or' pipes '|')}

\item{ignore_duplicate_filenames}{remove duplicated pdf filenames, e.g. /downloads/doc1.pdf and documents/doc1.pdf would return only one file.}

\item{output_form}{one of c("doc_paths","doc_text","matching_pages","text_lines") returns the contents of the pdfs rather than their names}

\item{reread}{the function saves an RDS of the pdf text in the search folder, do you want to delete the saved RDS file and reread the pdfs}
}
\value{
a list of PDF path locations that satisfy the regex (or alternatively a list of page numbers, the full text from the documents, or text from the specific lines where the regex text was detected)
}
\description{
Returns the names (or the internal text) of all PDF's that match your input regular expressions.
}
\details{
Note that the user function must have write access to the pdf folder location being searched
to save an RDS file of the text within the pdfs.
}
\examples{
\dontrun{ pdf_search_txt(c('1099', '(?i)corrected'), here('search_folder')) }
}
