% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_pdf_txt.R
\name{search_pdf_txt}
\alias{search_pdf_txt}
\title{Search a folder with PDFs with successive regular expressions}
\usage{
search_pdf_txt(
  search_regexes = c("", ""),
  pdf_folder = here::here(),
  reject_regex = NULL,
  include_pathnames = T,
  recursive = F,
  open_folder = F,
  return_pdf_text = F,
  ocr = F,
  reread = F,
  ignore_locns_with = NULL
)
}
\arguments{
\item{search_regexes}{vector of regexes to filter down the pdfs you're searching through}

\item{pdf_folder}{location of the folder with pdfs you want to read (uses here())}

\item{reject_regex}{single regex to reject any files that match the regex (broaden with regex 'or' pipes '|')}

\item{recursive}{search in folders as well as sub-folders}

\item{open_folder}{open folder locations in folder viewer}

\item{return_pdf_text}{return the text contents of the pdfs rather than their names}

\item{ocr}{read pdfs via optical character recognition rather than embedded pdf text}

\item{reread}{function saves an RDS of the pdf text in the folder, do you want to delete the RDS file and reread the pdfs}
}
\value{
a list of PDF locations that satisfy the regex
}
\description{
function must have write access to the pdf folder location to save an RDS file of the text within the pdfs
}
\examples{
\dontrun{ search_pdf_txt(c('1099', '(?i)corrected'), here('PDF_folder')) }
}
